Move wholesale tests to calculation tests
